[feat] MemberServiceImpl - registerMember() 완성

+++ src/main/java/com/sweetievegan/domain/service/MemberService.java
@@ -2,13 +2,12 @@ package com.sweetievegan.domain.service;
 
 import com.sweetievegan.domain.dto.MemberDto;
 
-import java.lang.reflect.Member;
 import java.util.List;
 
 public interface MemberService {
     List<MemberDto> getAllMembers();
     MemberDto findMemberByLoginId(String loginId);
-    int registerMember(MemberDto memberDto);
+    Long registerMember(MemberDto memberDto);
     MemberDto updateMemberDetail(Long memberId, MemberDto memberDto);
     int removeMember(Long memberId);
 }
